GH-2 :: simplify the code by eliminating the NavigationService - not needed anymore, because with the new ContentRetrieval, we no longer need to use DI to retrieve content in GetViewModel methods. We only need to inject the ContentRetrieval in the NavigationViewComponent
