--- src/Kickstart.Web/Features/Navigation/INavigationService.cs
@@ -1,8 +1,0 @@
-namespace Kickstart.Web.Features.Navigation;
-
-public interface INavigationService
-{
-    NavigationItemViewModel GetNavigationItemViewModel(NavigationItem navigationItem);
-
-    NavigationMenuViewModel GetNavigationMenuViewModel(NavigationMenu navigationMenu);
-}
